--- telegraf-kafka/appserver/static/components/togglepanel/togglepanel.css
@@ -1,45 +1,0 @@
-/*
- * Copyright (c) 2016-2017, OctoInsight Inc., All rights reserved.
- * Authored by Ryan Thibodeaux
- * see included LICENSE file (BSD 3-clause)
- */
-
-.toggledopen {
-  font-family: Arial, Helvetica, Arial, sans-serif;
-  padding: 9px 6px 6px 6px;
-  cursor: pointer;
-  display: inline;
-  float: left;
-  font-size: 20px;
-  font-weight: bolder;
-  -webkit-border-radius: 1px;
-  -moz-border-radius: 1px;
-  border-radius: 1px;
-}
-
-.toggledopen:before {
-  color: #551A8B;
-  content: "\25B2";
-}
-
-.toggledclosed {
-  font-family: Arial, Helvetica, Arial, sans-serif;
-  padding: 9px 6px 6px 6px;
-  cursor: pointer;
-  display: inline;
-  float: left;
-  font-size: 20px;
-  font-weight: bolder;
-  -webkit-border-radius: 1px;
-  -moz-border-radius: 1px;
-  border-radius: 1px;
-}
-
-.toggledclosed:before {
-  color: #551A8B;
-  content: "\25Bc";
-}
-
-.togglepanel-hidden {
-  display: none !important;
-}
