--- telegraf-kafka/appserver/static/telegraf-kafka-alerting.css
@@ -1,6 +1,0 @@
-
-/* Prevents green disabled button */
-.btn-primary[disabled] {
-    background-color: #337ab7;
-    opacity: .40;
-}
