+++ telegraf-kafka/default/app.conf
@@ -16,4 +16,4 @@ label = Kafka Smart Monitoring
 [launcher]
 author = Guilhem Marchand
 description = Kafka Smart Monitoring provides advanced and fast monitoring of Kafka and Confluent streaming platform
-version = 1.1.25
+version = 1.1.26
