+++ telegraf-kafka/default/data/ui/html/telegraf-kafka-configuration.html
@@ -7,11 +7,10 @@
     <link rel="shortcut icon" href="/en-US/static/@1143C71DB746E980E9A5DE0FB26F88D986BEC70C9C9F895CB9918539A4154D25.283/img/favicon.ico" />
     <link rel="stylesheet" type="text/css" href="{{SPLUNKWEB_URL_PREFIX}}/static/build/css/bootstrap-enterprise.css" />
     <link rel="stylesheet" type="text/css" href="{{SPLUNKWEB_URL_PREFIX}}/static/build/css/splunkjs-dashboard.css" />
-    <link rel="stylesheet" type="text/css" media="all" href="{{SPLUNKWEB_URL_PREFIX}}/static/app/telegraf-kafka/table_icon.css" />
+    <link rel="stylesheet" type="text/css" media="all" href="{{SPLUNKWEB_URL_PREFIX}}/static/app/telegraf-kafka/table.css" />
     <link rel="stylesheet" type="text/css" media="all" href="{{SPLUNKWEB_URL_PREFIX}}/static/app/telegraf-kafka/boottsrap-btn.css" />
     <link rel="stylesheet" type="text/css" media="all" href="{{SPLUNKWEB_URL_PREFIX}}/static/app/telegraf-kafka/hide_footer.css" />
     <link rel="stylesheet" type="text/css" media="all" href="{{SPLUNKWEB_URL_PREFIX}}/static/app/telegraf-kafka/common.css" />
-
 
         <meta name="referrer" content="never" />
         <meta name="referrer" content="no-referrer" />
@@ -398,7 +397,7 @@ var search_kafka_components = new SearchManager({
             "status_buckets": 0,
             "earliest_time": "-24h@h",
             "cancelOnUnload": true,
-            "search": "| inputlookup telegraf_kafka_configuration | eval keyid=_key, range=if(state=\"enabled\", 0, 1) | rangemap field=range low=0-0 severe=1-1 | fields keyid, label, state, range",
+            "search": "| inputlookup telegraf_kafka_configuration | eval keyid=_key | eval icon = \"icon|\" + case(state==\"enabled\",\"ico_good ico_small|icon-check|Up\", state!=\"enabled\", \"ico_error ico_small|icon-close|Down\") | fields keyid, label, state, icon | rename icon as \" \"",
             "sample_ratio": 1,
             "app": utils.getCurrentApp(),
             "auto_cancel": 90,
@@ -440,7 +439,6 @@ var search_kafka_components = new SearchManager({
             "count": 100,
             "dataOverlayMode": "none",
             "drilldown": "row",
-            "fields": "label,state,range",
             "percentagesRow": "false",
             "rowNumbers": "false",
             "totalsRow": "false",
@@ -823,6 +821,6 @@ var search_kafka_components = new SearchManager({
 );
 // ]]>
 </script>
-<script src="{{SPLUNKWEB_URL_PREFIX}}/static/app/telegraf-kafka/telegraf-kafka-configuration-table-icon.js" type="text/javascript"></script>
+<script src="{{SPLUNKWEB_URL_PREFIX}}/static/app/telegraf-kafka/table.js" type="text/javascript"></script>
 </body>
 </html>
