+++ telegraf-kafka/default/data/ui/views/overview_brokers.xml
@@ -1,4 +1,4 @@
-<form stylesheet="telegraf-kafka:common.css" hideEdit="True" isVisible="True">
+<form stylesheet="telegraf-kafka:common.css, telegraf-kafka:table.css" script="telegraf-kafka:table.js" hideEdit="True" isVisible="True">
     <label>Brokers</label>
     <search id="main_populate">
         <query>| savedsearch telegraf-kafka-kafka-brokers</query>
@@ -84,7 +84,14 @@
 | eval KBytesInPerSec=round(BytesInPerSec/1024, 2), KBytesOutPerSec=round(BytesOutPerSec/1024, 2)
 | rex field=jolokia_agent_url "//(?&lt;kafka_broker&gt;[^:]*)\:"
 | fields env, label, _time, kafka_broker, jolokia_agent_url, KBytesInPerSec, KBytesOutPerSec
-| stats latest(KBytesInPerSec) as "KBytes in", sparkline(avg(KBytesInPerSec),) As sparkline_in, latest(KBytesOutPerSec) as "KBytes out", sparkline(avg(KBytesOutPerSec),) As sparkline_out by env, label, kafka_broker, jolokia_agent_url</query>
+| stats latest(KBytesInPerSec) as "KBytes in", sparkline(avg(KBytesInPerSec),) As sparkline_in, latest(KBytesOutPerSec) as "KBytes out", sparkline(avg(KBytesOutPerSec),) As sparkline_out by env, label, kafka_broker, jolokia_agent_url
+| join type=outer env, label, kafka_broker [ | mstats latest(_value) as value where `telegraf_kafka_index` metric_name="kafka_controller.OfflinePartitionsCount.Value" OR metric_name="kafka_replica_manager.UnderReplicatedPartitions.Value" $env$ $label$ $jolokia_agent_url$ earliest=-2m latest=now by env, label, jolokia_agent_url, metric_name
+| eval {metric_name}=value
+| stats first(kafka_controller.OfflinePartitionsCount.Value) as offline, first(kafka_replica_manager.UnderReplicatedPartitions.Value) as under-replicated by env, label, jolokia_agent_url | rex field=jolokia_agent_url "//(?&lt;kafka_broker&gt;[^:]*)\:"
+| eval icon = "icon|" + if(offline=="0" AND 'under-replicated'=="0", "ico_good ico_small|icon-check|Up", "ico_error ico_small|icon-close|Down") ]
+| fields env, label, kafka_broker, "KBytes in", "sparkline_in", "KBytes out", "sparkline_out", offline, "under-replicated", icon
+| eval icon=if(isnull(icon), "icon|" + "ico_error ico_small|icon-close|Down", icon) | rename icon as " "
+| fillnull value="ERROR" offline, "under-replicated"</query>
                     <earliest>-60m@m</earliest>
                     <latest>now</latest>
                     <sampleRatio>1</sampleRatio>
