+++ telegraf-kafka/default/data/ui/views/telegraf-kafka-kafka-brokers.xml
@@ -1,4 +1,4 @@
-<form stylesheet="telegraf-kafka:common.css" hideEdit="true" isVisible="True">
+<form stylesheet="telegraf-kafka:common.css,telegraf-kafka:table.css" script="telegraf-kafka:table.js" hideEdit="true" isVisible="True">
     <label>Kafka Broker - Entity health view</label>
 
     <search id="broker_activity_summary">
@@ -1051,7 +1051,8 @@
 | eval {metric_name}=value
 | stats first(kafka_partition.*) as "*" by jolokia_agent_url, topic, partition
 | fields jolokia_agent_url, topic, partition, LogStartOffset, LogEndOffset, NumLogSegments, Size, UnderReplicatedPartitions
-| sort limit=0 topic, partition</query>
+| sort limit=0 topic, partition
+| eval icon = "icon|" + if('UnderReplicatedPartitions'=="0", "ico_good ico_small|icon-check|Up", "ico_error ico_small|icon-close|Down") | rename icon as " "</query>
                     <earliest>-1h</earliest>
                     <latest>now</latest>
                     <sampleRatio>1</sampleRatio>
